api/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
APScheduler==3.10.4
supabase==2.3.0
pandas==2.1.3 